move activateKeytarApis to ProfilesCache with dependent functions
