+++ packages/zowe-explorer-api/src/profiles/ProfilesCache.ts
@@ -10,12 +10,14 @@
  */
 
 import * as imperative from "@zowe/imperative";
+import { ZoweExplorerApi } from "./ZoweExplorerApi";
+import * as vscode from "vscode";
 import * as path from "path";
 import * as os from "os";
 import * as fs from "fs";
 import { URL } from "url";
-
-import { ZoweExplorerApi } from "./ZoweExplorerApi";
+import { KeytarCredentialManager } from "./KeytarCredentialManager";
+import { getSecurityModules } from "./CoreUtils";
 
 // TODO: find a home for constants
 export const CONTEXT_PREFIX = "_";
@@ -206,6 +208,23 @@ export class ProfilesCache {
             }
         }
         return baseProfile;
+    }
+
+    public async activateKeytarApis(initialized: boolean, isTheia: boolean): Promise<void> {
+        const scsActive = this.isSecureCredentialPluginActive();
+        if (scsActive) {
+            const keytar = getSecurityModules("keytar", isTheia);
+            if (!initialized && keytar) {
+                // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                KeytarCredentialManager.keytar = keytar;
+                const service: string = vscode.workspace.getConfiguration().get("Zowe Security: Credential Key");
+                await imperative.CredentialManagerFactory.initialize({
+                    service: service || "Zowe-Plugin",
+                    Manager: KeytarCredentialManager,
+                    displayName: "Zowe Explorer",
+                });
+            }
+        }
     }
 
     public isSecureCredentialPluginActive(): boolean {
