--- packages/zowe-explorer/i18n/sample/src/ZoweExplorerExtender.i18n.json
@@ -1,3 +1,0 @@
-{
-  "displayName": "Zowe Explorer"
-}
