+++ packages/zowe-explorer/i18n/sample/src/extension.i18n.json
@@ -1,6 +1,5 @@
 {
   "initialize.log.debug": "Initialized logger from VSCode extension",
-  "displayName": "Zowe Explorer",
   "initialize.log.error": "Error encountered while activating and initializing logger! ",
   "onDidSaveTextDocument1": "File was saved -- determining whether the file is a USS file or Data set.\n Comparing (case insensitive) ",
   "onDidSaveTextDocument2": " against directory ",
